Flytter dbkall til repository
