+++ src/main/kotlin/no/nav/k9/los/nyoppgavestyring/infrastruktur/abac/cache/PepCacheService.kt
@@ -5,10 +5,8 @@ import io.opentelemetry.extension.kotlin.asContextElement
 import io.opentelemetry.instrumentation.annotations.WithSpan
 import kotlinx.coroutines.*
 import kotliquery.TransactionalSession
-import kotliquery.queryOf
 import no.nav.k9.los.nyoppgavestyring.infrastruktur.abac.IPepClient
 import no.nav.k9.los.nyoppgavestyring.infrastruktur.db.TransactionalManager
-import no.nav.k9.los.nyoppgavestyring.infrastruktur.db.util.InClauseHjelper
 import no.nav.k9.los.nyoppgavestyring.mottak.oppgave.Oppgavestatus
 import no.nav.sif.abac.kontrakt.abac.Diskresjonskode
 import java.time.Duration
@@ -31,7 +29,7 @@ class PepCacheService(
     ) {
         transactionalManager.transaction { tx ->
             runBlocking(Dispatchers.IO) {
-                val oppgaverSomMåOppdateres = hentOppgaverMedStatusOgPepCacheEldreEnn(
+                val oppgaverSomMåOppdateres = pepCacheRepository.hentOppgaverMedStatusOgPepCacheEldreEnn(
                     tidspunkt = LocalDateTime.now() - gyldighet,
                     antall = 1,
                     status = status,
@@ -40,48 +38,6 @@ class PepCacheService(
                 oppgaverSomMåOppdateres.forEach { oppgave -> oppdater(tx, oppgave) }
             }
         }
-    }
-
-    private fun hentOppgaverMedStatusOgPepCacheEldreEnn(
-        tidspunkt: LocalDateTime = LocalDateTime.now(),
-        antall: Int = 1,
-        status: Set<Oppgavestatus>,
-        tx: TransactionalSession
-    ): List<PepCacheInput> {
-        val statusParametre = InClauseHjelper.tilParameternavn(status, "status")
-        val query = """
-                    SELECT o.oppgave_ekstern_id, 
-                    (select ov.verdi from oppgavefelt_verdi_part ov where ov.oppgave_id = o.id AND ov.feltdefinisjon_ekstern_id = 'saksnummer' AND ov.oppgavestatus IN ($statusParametre)) as saksnummer,
-                    (select ov.verdi from oppgavefelt_verdi_part ov where ov.oppgave_id = o.id AND ov.feltdefinisjon_ekstern_id = 'aktorId' AND ov.oppgavestatus IN ($statusParametre)) as aktor_id,
-                    (select ov.verdi from oppgavefelt_verdi_part ov where ov.oppgave_id = o.id AND ov.feltdefinisjon_ekstern_id = 'pleietrengendeAktorId' AND ov.oppgavestatus IN ($statusParametre)) as pleietrengende_aktor_id,
-                    (select ov.verdi from oppgavefelt_verdi_part ov where ov.oppgave_id = o.id AND ov.feltdefinisjon_ekstern_id = 'relatertPartAktorid' AND ov.oppgavestatus IN ($statusParametre)) as relatert_part_aktor_id
-                    FROM oppgave_v3_part o 
-                    LEFT JOIN OPPGAVE_PEP_CACHE opc ON o.oppgave_ekstern_id = opc.ekstern_id
-                    WHERE o.oppgavestatus IN ($statusParametre)
-                    AND (opc.oppdatert is null OR opc.oppdatert < :grense)
-                    ORDER BY opc.oppdatert NULLS FIRST
-                    LIMIT :limit
-                """.trimIndent()
-        return tx.run(
-            queryOf(
-                query,
-                buildMap {
-                    put("grense", tidspunkt)
-                    put("limit", antall)
-                    putAll(InClauseHjelper.parameternavnTilVerdierMap(status.map { it.kode }, "status"))
-                }
-            ).map { row ->
-                PepCacheInput(
-                    row.string("oppgave_ekstern_id"),
-                    row.stringOrNull("saksnummer"),
-                    listOfNotNull(
-                        row.stringOrNull("aktor_id"),
-                        row.stringOrNull("pleietrengende_aktor_id"),
-                        row.stringOrNull("relatert_part_aktor_id")
-                    )
-                )
-            }.asList
-        )
     }
 
     suspend fun oppdater(tx: TransactionalSession, oppgave: PepCacheInput): PepCache {
